+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2023-10-21
+### Changed
+- `@rarimo/rarime`:
+  - update circuit urls
+
+### Fixed
+- `@rarimo/rarime`:
+  - add `issuerDid` to filter credentials by issuer
+
+
 ## [0.6.0] - 2023-10-19
 ### Changed
 - `@rarimo/rarime`: